UPSTREAM: clk: meson: clk-pll: drop hard-coded rates from pll tables
Putting hard-coded rates inside the parameter tables assumes that
the parent is known and will never change. That's a big assumption
we should not make.

We have everything we need to recalculate the output rate using
the parent rate and the rest of the parameters. Let's do so and
drop the rates from the tables.

Acked-by: Neil Armstrong <narmstrong@baylibre.com>
Reviewed-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>
Tested-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>
Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>
(cherry picked from commit dd601dbc011e4eeda9e4c8c19ffe5b2fb33223dc)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>

Change-Id: Iedeaa5cc498a6bb2b2f1a4ae7b6631e13d9f2a3d