clk: fix clk_mux_val_to_index() error value
clk_mux_val_to_index() is meant to be used by .get_parent(), which
returns a u8, so when the value provided does not map to any valid index,
it is not a good idea to return a negative error value.

Instead, return num_parents which we know is an invalid index and let
CCF deal with it.

Fixes: 77deb66d262f ("clk: mux: add helper function for index/value translation")
Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>