WIP: mmc: meson-gx: fix irq ack
While cleaning the ISR of the meson-gx and acking only raised irqs,
the ack of the irq was moved at the very last stage of the function.

This was stable during the initial but it trigger issues with hs200,
under specific loads (booting android). Acking the irqs after calling
the mmc_request_done() causes the problem.

Moving the ack back to original place solves the issue. Since the irq
is edge triggered, it does not hurt to ack irq even earlier, so do
it early in the ISR.

Fixes: 9c5fdb07a28d ("mmc: meson-gx: ack only raised irq")
Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>