arm64: dts: meson-gxm-rbox-pro: add keep-power-in-suspend property in SDIO node
The WiFi firmware requires that the power is kept enabled while in
suspend mode. Add the keep-power-in-suspend property in the SDIO node
to specify that the power must be kept when entering in a system wide
suspend state.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>
Signed-off-by: Kevin Hilman <khilman@baylibre.com>
(cherry picked from commit e326c96778d23fbaa8c847f67ae84d113c2c119d)